Add Capabilities method.
In Docker 1.12, they added a Capabilities method to the Volume API that
designates whether or not something is "local" or "global." This is very
much a "local" plugin, which is the default.

This change doesn't actually do anything, since it just returns the
default. It just allows the plugin to be build again, because the helper
I rely on require this method be implemented.

For that reason, I am not doing another build with this commit.